requirements.txt
pyyaml==5.4.1
azure-identity==1.6.1
azure-keyvault==4.1.0
pandas==1.3.3